rename project to SynthesisRPGLoot
