+++ SynthesisRPGLoot.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HalgarisRPGLoot", "HalgarisRPGLoot\HalgarisRPGLoot.csproj", "{E3C85A2B-4E6D-4B5D-9E8D-810116F644AB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SynthesisRPGLoot", "SynthesisRPGLoot\SynthesisRPGLoot.csproj", "{E3C85A2B-4E6D-4B5D-9E8D-810116F644AB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
