+++ SynthesisRPGLoot/SynthesisMeta.json
