+++ SynthesisRPGLoot/SynthesisRPGLoot.csproj
